refactor: resolver structure
